feat(shop-the-look): Refactor page into smaller components
This commit completes Phase 2.5 of the Shop the Look refactoring plan.

The monolithic `pages/shop_the_look.py` has been broken down into smaller, more manageable pieces to improve maintainability and prepare for the next phase of refactoring.

- The main results UI is now in `components/shop_the_look/results_display.py`.
- The core event handlers (`on_click_vto_look`, `on_click_veo`, etc.) have been moved to `models/shop_the_look_handlers.py`.
- The main `pages/shop_the_look.py` file now primarily orchestrates the UI components.
- The global `header` component has been simplified to remove the status display, which will now be handled contextually within pages.

This commit also includes fixes for several regressions identified during the refactoring process, including circular imports and incorrect component usage.
